requirements.txt
absl-py==2.1.0
altgraph==0.17.4
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==24.2.0
black==24.8.0
blinker==1.7.0
cachetools==5.3.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.11
dataclasses==0.6
distlib==0.3.8
filelock==3.15.4
fire==0.6.0
Flask==3.0.2
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.53.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
google-api-core==2.17.1
google-api-python-client==2.139.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.65.4
h5py==3.11.0
httplib2==0.22.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.6
imageio==2.34.2
imantics==0.1.12
imgaug==0.4.0
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.4.1
kiwisolver==1.4.5
labelme2coco==0.2.6
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lxml==5.2.2
macholib==1.16.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.19.0
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
pixellib==0.7.1
platformdirs==4.2.2
pooch==1.8.2
portalocker==2.10.1
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.6
pydot==3.0.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyMatting==1.1.12
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rembg==2.0.58
requests==2.32.3
requests-oauthlib==1.4.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
sahi==0.11.18
scikit-image==0.24.0
scipy==1.14.0
setuptools==72.1.0
shapely==2.0.5
six==1.16.0
sympy==1.13.2
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
terminaltables==3.1.10
tifffile==2024.7.24
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.26.3
Werkzeug==3.0.1
wheel==0.44.0
wrapt==1.16.0
xmljson==0.2.1
yacs==0.1.8
